InjectionFrameWork/requirements.txt
pandas
psycopg2
pyodbc
azure-storage-blob
simple-salesforce
